requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pyyaml>=6.0.1
gspread>=6.0.0
google-auth>=2.29.0
python-dotenv>=1.0.1
keyring>=24.3.1
